+++ internal/sdkprovider/userconfig/service/clickhouse.go
@@ -51,7 +51,7 @@ func clickhouseUserConfig() *schema.Schema {
 				}},
 				MaxItems: 1024,
 				Optional: true,
-				Type:     schema.TypeList,
+				Type:     schema.TypeSet,
 			},
 			"ip_filter_string": {
 				Description: "Allow incoming connections from CIDR address block, e.g. '10.20.0.0/16'.",